Refactor to use `matchVersion`
