+++ test/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './match-version.js';
 export * from './snapshots.js';
 export * from './stubs/index.js';
 export * from './test-logger.js';
